+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - `set_cache_control_headers` to set `Cache-Control` headers for unpublished records
+- console "CurrentUser" controller allowing users to change their e-mail and password
 
 ### Changed
 - console `preview_url_for` - now defined in `Folio::Console::PreviewUrlFor` and expandable via `Rails.application.config.folio_console_preview_url_for_procs`